+++ src/views/toDoModal/repeatingEvent.vue
@@ -1,26 +1,35 @@
 <template>
-  <div id="reDropDown" class="header-menu-icons" type="button" data-bs-toggle="dropdown" data-bs-auto-close="outside"
-    :title="$t('ui.recurringTasks')">
+  <div
+    id="reDropDown"
+    class="header-menu-icons"
+    type="button"
+    data-bs-toggle="dropdown"
+    data-bs-auto-close="outside"
+    :title="$t('ui.recurringTasks')"
+  >
     <i id="btnRepeatingEvent" :class="{ 'bi-arrow-clockwise': !repeatingEvent, 'bi-arrow-repeat': repeatingEvent }"></i>
   </div>
 
   <ul class="dropdown-menu repeating-event-dropdown" aria-labelledby="btnRepeatingEvent">
     <div class="mx-3 d-flex flex-column drop-container">
-      <select class="form-select re-input" aria-label="Default select example" v-model="repeatingType"
-        :disabled="repeatingEvent">
+      <select
+        class="form-select re-input"
+        aria-label="Default select example"
+        v-model="repeatingType"
+        :disabled="repeatingEvent"
+      >
         <option value="">{{ $t("todoDetails.noRepeat") }}</option>
         <option value="3">{{ $t("todoDetails.daily") }}</option>
-        <option value="2">{{ $t("todoDetails.weekly") }} </option>
-        <option value="4">{{ $t("todoDetails.weekdays") }} </option>
-        <option value="5"> {{ $t("todoDetails.customWeekdays") }} </option>
+        <option value="2">{{ $t("todoDetails.weekly") }}</option>
+        <option value="4">{{ $t("todoDetails.weekdays") }}</option>
+        <option value="5">{{ $t("todoDetails.customWeekdays") }}</option>
         <option value="1">{{ $t("todoDetails.monthly") }}</option>
+        <option value="6">{{ $t("todoDetails.daysOfMonth") }}</option>
         <option value="0">{{ $t("todoDetails.yearly") }}</option>
       </select>
 
       <div v-if="repeatingType" class="d-flex flex-column">
-
-        <div v-show="repeatingType == '5' || (repeatingEvent && repeatingType == '2')"
-          class="weekDays-selector mt-3 mb-1">
+        <div v-show="repeatingType == '5' || (repeatingEvent && repeatingType == '2')" class="weekDays-selector mt-3 mb-1">
           <input type="checkbox" :disabled="repeatingEvent" id="weekday-mon" class="weekday" v-model="weekdays.mon" />
           <label for="weekday-mon">{{ moments().isoWeekday(1).locale(language).format("dd")[0] }}</label>
           <input type="checkbox" :disabled="repeatingEvent" id="weekday-tue" class="weekday" v-model="weekdays.tue" />
@@ -37,14 +46,27 @@
           <label for="weekday-sun">{{ moments().isoWeekday(7).locale(language).format("dd")[0] }}</label>
         </div>
 
+        <div v-show="repeatingType == '6'" class="mt-3 mb-1">
+          <input
+            v-model="daysOfMonth"
+            type="text"
+            class="form-control day-of-month"
+            :disabled="repeatingEvent"
+            placeholder="1,2,15"
+          />
+        </div>
 
         <div class="d-flex align-items-center justify-content-between re-input">
           <label class="opacity-50">{{ $t("todoDetails.interval") }}</label>
           <input type="number" class="form-control lex-shrink-1 counter" v-model="interval" :disabled="repeatingEvent" />
         </div>
 
-        <select class="form-select re-input" aria-label="Default select example" v-model="ocurrencesType"
-          :disabled="repeatingEvent">
+        <select
+          class="form-select re-input"
+          aria-label="Default select example"
+          v-model="ocurrencesType"
+          :disabled="repeatingEvent"
+        >
           <option value="">{{ $t("todoDetails.indefinitely") }}</option>
           <option value="ocurrences">
             {{ $t("todoDetails.occurrences") }}
@@ -52,10 +74,20 @@
           <option value="untilDate">{{ $t("todoDetails.untilDate") }}</option>
         </select>
 
-        <input v-if="ocurrencesType == 'ocurrences'" type="number" class="form-control re-input last-input"
-          v-model="ocurrences" :disabled="repeatingEvent" />
-        <input v-if="ocurrencesType == 'untilDate'" type="date" class="form-control re-input last-input"
-          v-model="untilDate" :disabled="repeatingEvent" />
+        <input
+          v-if="ocurrencesType == 'ocurrences'"
+          type="number"
+          class="form-control re-input last-input"
+          v-model="ocurrences"
+          :disabled="repeatingEvent"
+        />
+        <input
+          v-if="ocurrencesType == 'untilDate'"
+          type="date"
+          class="form-control re-input last-input"
+          v-model="untilDate"
+          :disabled="repeatingEvent"
+        />
       </div>
 
       <div class="d-flex flex-row re-form-row re-input">
@@ -88,6 +120,7 @@ export default {
       ocurrencesType: "",
       ocurrences: 1,
       untilDate: null,
+      daysOfMonth: "",
       weekdays: { mon: false, tue: false, wed: false, thu: false, fri: false, sat: false, sun: false },
     };
   },
@@ -156,6 +189,12 @@ export default {
         ruleOptions.freq = 2;
       }
 
+      if (this.repeatingType == 6) {
+        const daysOfMonthArray = this.daysOfMonth.split(",").map((num) => parseInt(num));
+        ruleOptions.bymonthday = daysOfMonthArray;
+        ruleOptions.freq = 1;
+      }
+
       if (this.ocurrencesType == "ocurrences") {
         ruleOptions.count = this.ocurrences;
       } else if (this.ocurrencesType == "untilDate") {
@@ -190,7 +229,7 @@ export default {
     },
     moments: function (date) {
       return moment(date);
-    }
+    },
   },
   watch: {
     repeatingEvent: function (newVal) {
@@ -202,7 +241,9 @@ export default {
         this.interval = rule.options.interval;
         this.ocurrences = rule.options.count;
         this.ocurrencesType = re.ocurrencesType;
-        this.untilDate = rule.options.until ? rule.options.until.toLocaleDateString("en-GB").split("/").reverse().join("-") : null;
+        this.untilDate = rule.options.until
+          ? rule.options.until.toLocaleDateString("en-GB").split("/").reverse().join("-")
+          : null;
         if (rule.options.byweekday) {
           rule.options.byweekday.includes(0) && (this.weekdays.mon = true);
           rule.options.byweekday.includes(1) && (this.weekdays.tue = true);
@@ -211,11 +252,16 @@ export default {
           rule.options.byweekday.includes(4) && (this.weekdays.fri = true);
           rule.options.byweekday.includes(5) && (this.weekdays.sat = true);
           rule.options.byweekday.includes(6) && (this.weekdays.sun = true);
+          this.repeatingType = 5;
         }
-
+        if (rule.options.bymonthday.length > 0) {
+          this.repeatingType = 6;
+          this.daysOfMonth = rule.options.bymonthday.join(",");
+        }
       } else {
         this.repeatingType = "";
         this.ocurrencesType = "";
+        this.daysOfMonth = "";
         this.interval = 1;
         this.untilDate = "";
         this.ocurrences = null;
@@ -227,7 +273,7 @@ export default {
     language: function () {
       return this.$store.getters.config.language;
     },
-  }
+  },
 };
 </script>
 
@@ -283,6 +329,19 @@ export default {
   border-bottom: 1px solid #464647;
 }
 
+.dark-theme .day-of-month {
+  background-color: #15161e;
+  border: 1px solid #464647;
+}
+
+.day-of-month::placeholder {
+  color: #9b9b9b;
+}
+
+.dark-theme .day-of-month::placeholder {
+  color: #838383;
+}
+
 .weekDays-selector {
   display: flex;
   justify-content: center;
@@ -292,7 +351,7 @@ export default {
   display: none !important;
 }
 
-.weekDays-selector input[type=checkbox]+label {
+.weekDays-selector input[type="checkbox"] + label {
   display: inline-block;
   border-radius: 3px;
   background: #eaecef;
@@ -306,17 +365,17 @@ export default {
   text-transform: uppercase;
 }
 
-.dark-theme .weekDays-selector input[type=checkbox]+label {
+.dark-theme .weekDays-selector input[type="checkbox"] + label {
   background: #15161e;
   color: #ffffff;
 }
 
-.weekDays-selector input[type=checkbox]:disabled+label {
+.weekDays-selector input[type="checkbox"]:disabled + label {
   cursor: unset;
   opacity: 0.6;
 }
 
-.weekDays-selector input[type=checkbox]:checked+label {
+.weekDays-selector input[type="checkbox"]:checked + label {
   background: #5c5c5c;
   color: #ffffff;
 }