+++ changelog.md
@@ -3,6 +3,9 @@ Jun 3, 2023
 
 - Adding Print and Export PDF Feature
 - Adding new sidebar submenu
+- To-do detail view textarea with dynamic height
+
+
 ## v2.1.0
 Jun 3, 2023