+++ changelog.md
@@ -1,3 +1,16 @@
+## v2.2.0
+Feb 14, 2024
+
+- Adding Print and Export PDF Feature
+- Adding new sidebar submenu
+- To-do detail view textarea with dynamic height
+- Rebranding (made by Hallgraph)
+- Using Sentry for error handle
+- Fixing move old tasks to current day
+- Adding recurring task filter
+- Adding Month days recurring tasks
+
+
 ## v2.1.0
 Jun 3, 2023
 
@@ -8,7 +21,7 @@ Jun 3, 2023
 - Upgrading all tech stack libraries
 - Adding new repeating event option for custom weekdays selector
 - Improving splash screen
-- Adding sent subtask to bottom on clomplete setting
+- Adding sent subtask to bottom on complete setting
 - Todo Description markdown link now redirect to external link in web browser
 - Adding week start on monday setting
 - Improving Chinese and Russian languages